fixing issues for 64-bit machines

+++ macs/cmusdrg_macs/trunk/src/lib/cmac_framer.h
@@ -35,12 +35,12 @@ static const long POSTAMBLE_LEN=16;
 static const int MAX_FRAME_SIZE = 1500;
 
 typedef struct d_frame_hdr_t {
-  long src_addr;
-  long dst_addr;
-  long payload_len;
-  unsigned long seq_num;
+  uint32_t src_addr;
+  uint32_t dst_addr;
+  uint32_t payload_len;
+  uint32_t seq_num;
   bool ack;
-  long crc;
+  uint32_t crc;
 } __attribute__((__packed__));
 
 #endif
